+++ inappmessaging/src/main/java/io/karte/android/inappmessaging/InAppMessaging.kt
@@ -16,7 +16,6 @@
 package io.karte.android.inappmessaging
 
 import android.app.Activity
-import android.os.Build
 import android.os.Handler
 import android.os.Looper
 import android.view.Window
@@ -35,6 +34,7 @@ import io.karte.android.inappmessaging.internal.PanelWindowManager
 import io.karte.android.inappmessaging.internal.preview.PreviewParams
 import io.karte.android.tracking.Event
 import io.karte.android.tracking.MessageEvent
+import io.karte.android.tracking.MessageEventName
 import io.karte.android.tracking.MessageEventType
 import io.karte.android.tracking.Tracker
 import io.karte.android.tracking.client.TrackRequest
@@ -61,7 +61,7 @@ class InAppMessaging : Library, ActionModule, UserModule, TrackModule, ActivityL
         self = this
         app.application.registerActivityLifecycleCallbacks(this)
         this.app = app
-        this.processor = IAMProcessor(app.application, panelWindowManager)
+        this.processor = IAMProcessor(app.application, panelWindowManager, app.config.isAutoScreenBoundaryEnabled)
         this.config = app.libraryConfig(InAppMessagingConfig::class.java) ?: InAppMessagingConfig.build()
         app.register(this)
     }
@@ -126,6 +126,13 @@ class InAppMessaging : Library, ActionModule, UserModule, TrackModule, ActivityL
         get() = listOf(ExpiredMessageOpenEventRejectionFilterRule())
 
     override fun prepare(event: Event): Event {
+        when (event.eventName.value) {
+            MessageEventName.MessageReady.value,
+            MessageEventName.MessageOpen.value,
+            MessageEventName.MessageClick.value -> {
+                event.values.put("_is_auto_screen_boundary_enabled", app.config.isAutoScreenBoundaryEnabled)
+            }
+        }
         if (event.eventName.value == "view") {
             uiThreadHandler.post {
                 processor.handleView(event.values)
@@ -259,18 +266,16 @@ class InAppMessaging : Library, ActionModule, UserModule, TrackModule, ActivityL
     }
 
     private fun clearWebViewCookies() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            val cookieManager = CookieManager.getInstance()
-            val allCookies = cookieManager.getCookie(COOKIE_DOMAIN) ?: return
-            allCookies
-                .split("; ")
-                .filter { !it.isBlank() && it.contains("=") }
-                .forEach {
-                    val cookieString = it.substringBefore("=") + "=; Domain=" + COOKIE_DOMAIN
-                    cookieManager.setCookie(COOKIE_DOMAIN, cookieString)
-                }
-            cookieManager.flush()
-        }
+        val cookieManager = CookieManager.getInstance()
+        val allCookies = cookieManager.getCookie(COOKIE_DOMAIN) ?: return
+        allCookies
+            .split("; ")
+            .filter { !it.isBlank() && it.contains("=") }
+            .forEach {
+                val cookieString = it.substringBefore("=") + "=; Domain=" + COOKIE_DOMAIN
+                cookieManager.setCookie(COOKIE_DOMAIN, cookieString)
+            }
+        cookieManager.flush()
     }
 
     private fun trackMessageSuppressed(message: JSONObject, reason: String) {